data/postgres/pyproject.toml
[project]
name = "postgres"
version = "0.1.0"
description = ""
authors = [
    {name = "Dustin van Weersel",email = "dustin.vanweersel@foundationzero.org"}
]
readme = "README.md"
requires-python = ">=3.13"

[tool.poetry.dependencies]
pydantic-settings = "^2.7.1"
psycopg = {extras = ["binary"], version = "^3.2.3"}
python-dotenv = "^1.1.0"

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
